Added support for passwordless guest IRC login via loginserv. Loginserv will now understand messages lacking a password and assume an empty string for that password (e.g. guest account).
This was not working because of a quirk of the IRC PRIVMSG protocol:
multiword PRIVMSG are prefixed by a colon ":" character, while single
word ones (e.g. /msg loginserv guest) are not. The code expected msg
to loginserv to always be multiword (i.e. contain a user and pass str
separated by a space).