HLTracker, HLServer: HLTracker servers (for publishing) accept UDP packets, not TCP... Corrected the HLTracker client class to use UDP in place of TCP. Note that this has some drawbacks (we can't detect if the UDP packet has not made it to the host). The previous code was sending the tracker update to the TCP port listening for client connections, which wasn't updating anything (obviously).
config.py: Changed the default port in the trackerlist example from
5497 to 5499 (the correct UDP tracker port).

HLProtocol: After finding a published tracker packet structure I was
able to correctly implement the tracker update packet builder function.
The tracker packet should now be 100% up to specs. Tested it
successfuly, observed that adding Users to the server incremented the
count on the server, visible on a tracker client. With these changes
the tracker client functionality is finally ready.